-Wall cleanups; fix implicit decls and clobber unused vars
git-svn-id: svn://svn.berlios.de/lincity-ng/trunk@252 8226ee30-1aee-0310-a633-b68f58d1eb45

+++ readpng.c
@@ -112,7 +112,7 @@ load_png_graphics (char *txt_filename, char *png_filename)
     png_get_IHDR(png_ptr, info_ptr, &width, &height, &bit_depth, &color_type,
 	&interlace_type, NULL, NULL);
 
-    printf ("PNG Header: %d x %d, bd=%d, ct=%d\n", height, width,
+    printf ("PNG Header: %d x %d, bd=%d, ct=%d\n", (int)height, (int)width,
 	bit_depth, color_type);
 
     /* Set up the data transformations you want.  Note that these are all
